requirements.txt
openai-whisper
torch
torchaudio
sounddevice
soundfile
numpy
